+++ src/components/drawer/index.ts
@@ -1,1 +1,3 @@
 export * from './Drawer'
+export * from './DrawerFormFooter'
+export * from './DrawerHeader'