+++ src/pages/trackedEntityTypes/Edit.tsx
@@ -36,7 +36,7 @@ const fieldFilters = [
     'minAttributesRequiredToSearch',
     'maxTeiCountToReturn',
     'featureType',
-    'trackedEntityTypeAttributes[trackedEntityAttribute[id,displayName,unique],mandatory,searchable,displayInList]',
+    'trackedEntityTypeAttributes[trackedEntityAttribute[id,displayName,unique,valueType],mandatory,searchable,displayInList]',
 ] as const
 
 export type TrackedEntityTypeFormValues = PickWithFieldFilters<