+++ CHANGELOG.md
@@ -1,3 +1,97 @@
+# [0.136.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.135.0...v0.136.0) (2026-03-03)
+
+
+### Bug Fixes
+
+* tooltip with conditional text ([6fa2cea](https://github.com/dhis2/maintenance-app-beta/commit/6fa2cead5d554fe802a1307f6f02f790e0df6601))
+
+
+### Features
+
+* get tea value type for edit form ([5357e02](https://github.com/dhis2/maintenance-app-beta/commit/5357e029812d3521be72ba5b5703f7d93e832f06))
+
+# [0.135.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.134.0...v0.135.0) (2026-03-03)
+
+
+### Bug Fixes
+
+* add span with tab index back into drawer ([#784](https://github.com/dhis2/maintenance-app-beta/issues/784)) ([097b982](https://github.com/dhis2/maintenance-app-beta/commit/097b9822dd9dd887face0133e8b1214c75d4dde0))
+* make required labels clear ([#785](https://github.com/dhis2/maintenance-app-beta/issues/785)) ([bcdadc6](https://github.com/dhis2/maintenance-app-beta/commit/bcdadc6812d4ebdac6460ce380d5f535ef6f4019))
+
+
+### Features
+
+* [DHIS2-20784] Program rule form ([#752](https://github.com/dhis2/maintenance-app-beta/issues/752)) ([7a9c577](https://github.com/dhis2/maintenance-app-beta/commit/7a9c5770a5063c4f0ec131d56ec525503324dffe))
+
+# [0.134.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.133.1...v0.134.0) (2026-03-02)
+
+
+### Features
+
+* move programsWip to programs ([#782](https://github.com/dhis2/maintenance-app-beta/issues/782)) ([f3d7e80](https://github.com/dhis2/maintenance-app-beta/commit/f3d7e805dd2f29b2dc2bab6f9900683042a574fd))
+
+## [0.133.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.133.0...v0.133.1) (2026-03-02)
+
+
+### Bug Fixes
+
+* reapply non editing of event programs ([f6f6fed](https://github.com/dhis2/maintenance-app-beta/commit/f6f6fed47c8100c09ba3df0efc2570fe6a213682))
+
+# [0.133.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.132.0...v0.133.0) (2026-03-02)
+
+
+### Features
+
+* **drawer:** header/footer, globalShell conditional height, cleanup ([#773](https://github.com/dhis2/maintenance-app-beta/issues/773)) ([7528300](https://github.com/dhis2/maintenance-app-beta/commit/7528300336966dc5dd0d446212a39d6a747a3b3f))
+
+# [0.132.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.131.0...v0.132.0) (2026-02-26)
+
+
+### Features
+
+* add details fields ([#769](https://github.com/dhis2/maintenance-app-beta/issues/769)) ([31aa2e4](https://github.com/dhis2/maintenance-app-beta/commit/31aa2e4bdcccbc7ed4d4f300659c040366c9e2b7))
+
+# [0.131.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.130.1...v0.131.0) (2026-02-26)
+
+
+### Bug Fixes
+
+* value type logic [DHIS2-20962] ([#777](https://github.com/dhis2/maintenance-app-beta/issues/777)) ([475648e](https://github.com/dhis2/maintenance-app-beta/commit/475648e4415334fd42e3eadd062a4179043deeb7))
+
+
+### Features
+
+* add missing translations props [DHIS2-20971] ([db783cf](https://github.com/dhis2/maintenance-app-beta/commit/db783cf07499773bfd2cdeab01adb496590de7af))
+
+## [0.130.1](https://github.com/dhis2/maintenance-app-beta/compare/v0.130.0...v0.130.1) (2026-02-26)
+
+
+### Bug Fixes
+
+* allow org unit group and level deselection in data set forms ([51f2734](https://github.com/dhis2/maintenance-app-beta/commit/51f273438c2acb147d17ae6739c1987048174b9a))
+* preserve spaces in list elements ([#774](https://github.com/dhis2/maintenance-app-beta/issues/774)) ([20eb148](https://github.com/dhis2/maintenance-app-beta/commit/20eb148a86cea901f51d996adc9606aa8d95d8ab))
+
+# [0.130.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.129.0...v0.130.0) (2026-02-25)
+
+
+### Features
+
+* better distinction between program and program stages notification  ([edfd5d9](https://github.com/dhis2/maintenance-app-beta/commit/edfd5d9caa2782f36283ed3b57e7ff9d0778e786))
+
+# [0.129.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.128.0...v0.129.0) (2026-02-25)
+
+
+### Features
+
+* add missing field and hide edit of event programs ([#778](https://github.com/dhis2/maintenance-app-beta/issues/778)) ([7a9ad5b](https://github.com/dhis2/maintenance-app-beta/commit/7a9ad5b8153f2fff78ba3ac10dc952eb713fb19a))
+
+# [0.128.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.127.0...v0.128.0) (2026-02-23)
+
+
+### Features
+
+* implement program stage notifications form ([#771](https://github.com/dhis2/maintenance-app-beta/issues/771)) ([95770e3](https://github.com/dhis2/maintenance-app-beta/commit/95770e39ac4637fea00f29a5a79d8ae31e9246b3))
+
 # [0.127.0](https://github.com/dhis2/maintenance-app-beta/compare/v0.126.1...v0.127.0) (2026-02-19)